+++ css/base/styles.css
@@ -7,7 +7,7 @@
   --tph-highlight: #2eac68 !important;
   --tph-text-highlight: #2cad69 !important;
   --tph-button-hover: #2dd67b !important;
-  --placeholder: #c4c4c4 !important;
+  --placehol  der: #c4c4c4 !important;
   --input: #b9c8d3 !important;
 }
